+++ src/idc/imgaug/filter/__init__.py
@@ -7,6 +7,7 @@ from ._crop import Crop
 from ._crop_to_label import CropToLabel
 from ._enhance import Enhance, ENHANCEMENTS, ENHANCEMENT_CONTRAST, ENHANCEMENT_COLOR, ENHANCEMENT_SHARPNESS, ENHANCEMENT_BRIGHTNESS
 from ._find_contours import FindContours, CONNECTIVITY, CONNECTIVITY_LOW, CONNECTIVITY_HIGH
+from ._find_contours_cv2 import FindContoursCV2
 from ._flip import Flip
 from ._gaussian_blur import GaussianBlur
 from ._hsl_grayscale import HSLGrayScale